refactor: 모달 provider 폴더 위치 수정
