+++ src/providers/ModalProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import Modal from '../common/modals/Modal';
+import Modal from '../components/common/modals/Modal';
 
 export default function ModalProvider() {
   const [isMounted, setIsMounted] = useState(false);
